Switch ImgFile to using uint64_t explicitly
size_t and off_t are 32-bit values in Emscripten, which causes issues
with disk images larger than 4GB. Use the explicit type (which is more
consistent with the rest of the codebase anyway).

+++ utils/imgfile_sdl.cpp
@@ -46,20 +46,20 @@ void ImgFile::close()
     impl->stream.close();
 }
 
-size_t ImgFile::size() const
+uint64_t ImgFile::size() const
 {
     impl->stream.seekg(0, impl->stream.end);
     return impl->stream.tellg();
 }
 
-size_t ImgFile::read(void* buf, off_t offset, size_t length) const
+uint64_t ImgFile::read(void* buf, uint64_t offset, uint64_t length) const
 {
     impl->stream.seekg(offset, std::ios::beg);
     impl->stream.read((char *)buf, length);
     return impl->stream.gcount();
 }
 
-size_t ImgFile::write(const void* buf, off_t offset, size_t length)
+uint64_t ImgFile::write(const void* buf, uint64_t offset, uint64_t length)
 {
     impl->stream.seekg(offset, std::ios::beg);
     impl->stream.write((const char *)buf, length);
